+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Not released
 Features
 --------
 * Ensure the driver can connect when invalid peer hosts are in system.peers (PYTHON-1260)
+* Implement protocol v5 checksumming (PYTHON-1258)
 
 Bug Fixes
 ---------